refactor: Moved packages, updated build process
