+++ package.json
@@ -2,17 +2,17 @@
   "name": "react-colorful",
   "version": "2.2.1",
   "description": "A tiny color picker component for modern React apps",
-  "source": "src/packages/hex.tsx",
+  "source": "src/hex.tsx",
   "main": "dist/index.js",
   "module": "dist/index.module.js",
   "esmodule": "dist/index.esmodule.js",
   "umd:main": "dist/index.umd.js",
-  "types": "dist/packages/hex.d.ts",
+  "types": "dist/hex.d.ts",
   "scripts": {
     "lint": "eslint src/**/*.{ts,tsx} demo/src/**/*.js",
     "size": "npm run build && size-limit",
     "test": "jest tests",
-    "build": "node ./build-packages.js",
+    "build": "node --experimental-json-modules ./build-packages.mjs",
     "prepublishOnly": "npm run build",
     "start-demo": "parcel demo/src/index.html --out-dir demo/dist --open",
     "build-demo": "del-cli 'demo/dist/*' && parcel build demo/src/index.html --out-dir demo/dist --public-url /react-colorful/",
@@ -100,6 +100,7 @@
     "@types/react": "^16.9.46",
     "@typescript-eslint/eslint-plugin": "^3.9.1",
     "@typescript-eslint/parser": "^3.9.1",
+    "del": "^5.1.0",
     "del-cli": "^3.0.1",
     "eslint": "^7.5.0",
     "eslint-config-prettier": "^6.11.0",
