+++ src/HexInput.ts
@@ -1,0 +1,2 @@
+import HexInput from "./components/HexInput";
+export default HexInput;
