+++ src/hex.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 
-import ColorPicker from "../components/ColorPicker";
-import { ColorModel, ColorPickerBaseProps } from "../types";
-import { equalHex } from "../utils/compare";
-import { hexToHsv, hsvToHex } from "../utils/convert";
+import ColorPicker from "./components/ColorPicker";
+import { ColorModel, ColorPickerBaseProps } from "./types";
+import { equalHex } from "./utils/compare";
+import { hexToHsv, hsvToHex } from "./utils/convert";
 
 const colorModel: ColorModel<string> = {
   defaultColor: "000",
