--- src/packages/HexInput.ts
@@ -1,2 +1,0 @@
-import HexInput from "../components/HexInput";
-export default HexInput;
